pyproject.toml
[project]
name = "idftestproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "airflow-providers-clickhouse>=0.0.1",
    "apache-airflow==3.0.4",
    "asyncpg>=0.30.0",
    "clickhouse-connect>=0.8.18",
    "dbt-clickhouse>=1.9.2",
    "dbt-core>=1.10.10",
    "faker>=37.5.3",
    "kafka-python>=2.2.15",
    "psycopg2-binary>=2.9.10",
    "tenacity>=9.1.2",
]
